Refactor environment configuration using @t3-oss/env and remove shared package
