+++ packages/api/package.json
@@ -8,17 +8,18 @@
     "build": "tsc",
     "dev": "tsx --watch src/server.ts",
     "start": "node dist/server.js",
-    "env:check": "node ../shared/scripts/env-check.js",
+    "env:check": "node ./scripts/env-check.js",
     "lint": "biome check .",
     "format": "biome format --write ."
   },
   "dependencies": {
     "@hono-payload/db": "workspace:*",
-    "@hono-payload/shared": "workspace:*",
     "hono": "^4.6.8",
     "@hono/node-server": "^1.13.7",
     "cors": "^2.8.5",
-    "drizzle-orm": "0.44.2"
+    "drizzle-orm": "0.44.2",
+    "@t3-oss/env-core": "^0.12.0",
+    "dotenv-flow": "^4.1.0"
   },
   "devDependencies": {
     "@types/cors": "^2.8.17",
