+++ packages/api/src/server.ts
@@ -2,7 +2,7 @@ import { serve } from "@hono/node-server";
 import { Hono } from "hono";
 import { cors } from "hono/cors";
 import { logger } from "hono/logger";
-import { env } from "@hono-payload/shared/env";
+import { env } from "./env";
 import { prettyJSON } from "hono/pretty-json";
 
 // Import API routes
