+++ packages/cms/package.json
@@ -9,7 +9,7 @@
     "devsafe": "rm -rf .next && cross-env NODE_OPTIONS=--no-deprecation next dev",
     "generate:importmap": "cross-env NODE_OPTIONS=--no-deprecation payload generate:importmap",
     "generate:types": "cross-env NODE_OPTIONS=--no-deprecation payload generate:types",
-    "env:check": "node ../shared/scripts/env-check.js",
+    "env:check": "node ./scripts/env-check.js",
     "lint": "biome check .",
     "format": "biome format --write .",
     "payload": "cross-env NODE_OPTIONS=--no-deprecation payload",
@@ -17,7 +17,6 @@
   },
   "dependencies": {
     "@hono-payload/db": "workspace:*",
-    "@hono-payload/shared": "workspace:*",
     "@payloadcms/db-postgres": "3.52.0",
     "@payloadcms/next": "3.52.0",
     "@payloadcms/payload-cloud": "3.52.0",
@@ -29,6 +28,7 @@
     "graphql": "^16.8.1",
     "next": "15.4.4",
     "payload": "3.52.0",
+    "@t3-oss/env-nextjs": "^0.12.0",
     "react": "19.1.1",
     "sharp": "0.34.2",
     "uuid": "^11.1.0"
