+++ packages/db/package.json
@@ -26,14 +26,15 @@
     "db:migrate": "drizzle-kit migrate",
     "db:push": "drizzle-kit push",
     "db:studio": "drizzle-kit studio",
-    "env:check": "node ../shared/scripts/env-check.js",
+    "env:check": "node ./scripts/env-check.js",
     "lint": "biome check .",
     "format": "biome format --write ."
   },
   "dependencies": {
-    "@hono-payload/shared": "workspace:*",
     "drizzle-orm": "0.44.2",
-    "postgres": "^3.4.4"
+    "postgres": "^3.4.4",
+    "@t3-oss/env-core": "^0.12.0",
+    "dotenv-flow": "^4.1.0"
   },
   "devDependencies": {
     "@types/node": "^22.5.4",
