+++ packages/db/src/db/connection.ts
@@ -1,7 +1,7 @@
 import { drizzle } from "drizzle-orm/postgres-js";
 import postgres from "postgres";
 import * as schema from "./schema/index.js";
-import { env } from "@hono-payload/shared/env";
+import { env } from "../env";
 
 // Create PostgreSQL connection using validated env
 const connectionString = env.DATABASE_URI;
